Sim elevator working. sorta
The simulated elevator works by grabbing a value from constants and using that as an input. I am working on making it work off a command as usual.
also got the real elevator code more refined.
added the subsystem set minip for the coral minipulator. that is in the very early stages of working at this time.
next, I need to add a motor for the climber. as far as I know, its a winch, so I plan on making it use the opperator joystick.